+++ packages/ui/src/components/tabs/PublishAPI.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from 'react'
 import { jsx, Input, Flex, Button, Styled } from 'theme-ui'
 import { useCreateSubdomain } from '../../hooks/ens/useCreateSubdomain'
 import { useStateValue } from '../../state/state'
-import getMetaDataFromPackageHash from '../../utils/getMetaDataFromPackageHash'
+import getMetaDataFromPackageHash from '../../services/ipfs/getMetaDataFromPackageHash'
 import Card from '../Card'
 import Modal from '../Modal'
 import { MAIN_DOMAIN, ZERO_ADDRESS } from '../../constants'
